requirements.txt
Django==5.1
django-cleanup==8.1.0
django-environ==0.11.2
pillow==10.4.0
sorl-thumbnail==12.10.0
django-unfold==0.36.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
flake8==7.1.1
flake8-commas==4.0.0
flake8-dunder-all==0.4.1
flake8-import-order==0.18.2
flake8-quotes==3.4.0
pep8-naming==0.14.1
whitenoise==6.7.0
psycopg2-binary==2.9.9